% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiCopLikCV.R
\name{CondiCopLikCV}
\alias{CondiCopLikCV}
\title{Cross-validated likelihood.}
\usage{
CondiCopLikCV(u1, u2, family, X, xind = 100, degree = 1, eta, nu,
  kernel = KernEpa, band, optim_fun, cveta_out = FALSE, cl = NA)
}
\arguments{
\item{u1}{Vector of first uniform response.}

\item{u2}{Vector of second uniform response.}

\item{family}{An integer defining the bivariate copula family to use.  See \code{\link{ConvertPar}}.}

\item{X}{Vector of observed covariate values.}

\item{xind}{Vector of indices in \code{X} at which to calculate leave-one-out parameter estimates.  Can also be supplied as a single integer, in which case \code{xind} equally spaced observations are taken from \code{X}.}

\item{degree}{Integer specifying the polynomial order of the local likelihood function.  Currently only 0 and 1 are supported.}

\item{eta, nu, kernel, band, optim_fun, cl}{See \code{\link{CondiCopLocFit}}.}

\item{cveta_out}{If \code{TRUE}, the CV estimate of eta at each point in \code{X} in addition to the CV log-likelihood.}
}
\value{
Scalar value of the cross-validated log-likelihood, or a list with elements \code{x}, \code{eta}, \code{nu}, and \code{loglik} if \code{cveta_out = TRUE}, the first three being the same as for \code{\link{CondiCopLocFit}}.
}
\description{
Leave-one-out local likelihood copula parameter estimates are interpolated, then used to calculate the conditional copula likelihood function.
}
\seealso{
This function is typically used in conjunction with \code{\link{CondiCopSelect}}; see examples there.
}
