% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiCopLocFit1.R
\name{CondiCopLocFit1}
\alias{CondiCopLocFit1}
\title{Local likelihood estimation at a single covariate value.}
\usage{
CondiCopLocFit1(u1, u2, family, z, wgt, degree = c("linear", "constant"),
  eta, nu)
}
\arguments{
\item{u1}{Vector of first uniform response.}

\item{u2}{Vector of second uniform response.}

\item{family}{The copula family to use.  See \code{\link{ConvertPar}}.}

\item{z}{Vector of centered covariate values, i.e., \code{X-x}.}

\item{wgt}{Vector of kernel weights.}

\item{degree}{Character vector "constant" or "linear" specifying the type of local likelihood fit.}

\item{eta}{Initial value of the copula dependence parameter.}

\item{nu}{Value of other copula parameters (if they exist).}
}
\value{
List with elements \code{eta} and \code{loglik}.
}
\description{
Local likelihood estimation at a single covariate value.
}
