% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kernel}
\alias{kernel}
\alias{KernEpa}
\alias{KernGaus}
\alias{KernBeta}
\alias{KernBiQuad}
\alias{KernTriAng}
\title{Local likelihood kernel functions.}
\usage{
KernEpa(t)

KernGaus(t)

KernBeta(t, par)

KernBiQuad(t)

KernTriAng(t)
}
\arguments{
\item{t}{Vector of distances from mode of kernel.}

\item{par}{...}
}
\value{
Vector of kernel weights.
}
\description{
Local likelihood kernel functions.
}
\details{
Describe kernels here.
}
