% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernFun.R
\name{KernFun}
\alias{KernFun}
\alias{KernEpa}
\alias{KernGaus}
\alias{KernBeta}
\alias{KernBiQuad}
\alias{KernTriAng}
\title{Local likelihood kernel functions.}
\usage{
KernEpa(t)

KernGaus(t)

KernBeta(t, par = 0.5)

KernBiQuad(t)

KernTriAng(t)
}
\arguments{
\item{t}{Vector of distances from mode of kernel.}

\item{par}{Shape parameter for Beta kernel (positive scalar).}
}
\value{
Vector of kernel weights.
}
\description{
Local likelihood kernel functions.
}
\details{
Describe kernels here.
}
