% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalCop-package.R
\docType{package}
\name{LocalCop-package}
\alias{LocalCop}
\alias{LocalCop-package}
\title{LocalCop: Local Likelihood Inference for Conditional Copula Models}
\description{
Implements a local likelihood estimator for the dependence parameter in bivariate conditional copula models.  Copula family and local likelihood bandwidth parameters are selected by leave-one-out cross-validation.  The models are implemented in 'TMB', meaning that the local score function is efficiently calculated via automated differentiation (AD), such that quasi-Newton algorithms may be used for parameter estimation.
}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

Authors:
\itemize{
  \item Elif Acar \email{elif.acar@umanitoba.ca}
}

}
\keyword{internal}
