% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvCopLikCV.R
\name{SurvCopLikCV}
\alias{SurvCopLikCV}
\title{Cross-validated survival copula likelihood.}
\usage{
SurvCopLikCV(
  u1,
  u2,
  status1,
  status2,
  family,
  x,
  degree = 1,
  eta,
  kernel = KernEpa,
  band,
  optim_fun,
  cveta_out = FALSE,
  cv_all = FALSE,
  cl = NA
)
}
\arguments{
\item{u1}{Vector of first uniform response.}

\item{u2}{Vector of second uniform response.}

\item{status1}{Vector of censoring indicators for the first variable.}

\item{status2}{Vector of censoring indicators for the second variable.}

\item{family}{An integer defining the bivariate copula family to use.  See \code{\link[=ConvertPar]{ConvertPar()}}.}

\item{x}{Vector of observed covariate values.}

\item{degree}{Integer specifying the polynomial order of the local likelihood function.  Currently only 0 and 1 are supported.}

\item{eta, kernel, band, optim_fun, cl}{See \code{\link[=SurvCopLocFit]{SurvCopLocFit()}}.}

\item{cveta_out}{If \code{TRUE}, return the CV estimate of eta at each point in \code{x} in addition to the CV log-likelihood.}

\item{cv_all}{If \code{FALSE}, evaluate the CV likelihood at only the leave-one-out observations specified by \code{xind}.  Otherwise, interpolate the leave-one-out estimates of eta to all values in \code{x}, and evaluate the CV likelihood at all observations.}
}
\value{
If \code{cveta_out = FALSE}, scalar value of the cross-validated log-likelihood.  Otherwise, a list with elements:
\describe{
\item{\code{x}}{The sorted values of \code{x}.}
\item{\code{eta}}{The leave-one-out estimates interpolated from the values in \code{xind} to all of those in \code{x}.}
\item{\code{loglik}}{The cross-validated log-likelihood.}
}
}
\description{
Leave-one-out local likelihood copula parameter estimates are obtained and used to calculate the conditional copula likelihood function under censoring.
}
