% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsVine-package.R
\docType{package}
\name{tsVine-package}
\alias{tsVine}
\alias{tsVine-package}
\title{tsVine: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

Authors:
\itemize{
  \item Elif Acar \email{elif.acar@umanitoba.ca}
}

}
\keyword{internal}
