% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiCopLikCV.R
\name{CondiCopLikCV}
\alias{CondiCopLikCV}
\title{Cross-validated likelihood.}
\usage{
CondiCopLikCV(u1, u2, family, X, xind = 100, degree = 1, eta, nu,
  kernel = KernEpa, band, optim_fun, cveta_out = FALSE, cl = NA)
}
\arguments{
\item{u1}{Vector of first uniform response.}

\item{u2}{Vector of second uniform response.}

\item{family}{An integer defining the bivariate copula family to use.  See \code{\link{ConvertPar}}.}

\item{X}{Vector of observed covariate values.}

\item{xind}{Vector of indices in \code{X} at which to calculate leave-one-out parameter estimates.  Can also be supplied as a single integer, in which case \code{xind} equally spaced observations are taken from \code{X}.}

\item{degree}{Integer specifying the polynomial order of the local likelihood function.  Currently only 0 and 1 are supported.}

\item{eta, nu, kernel, band, optim_fun, cl}{See \code{\link{CondiCopLocFit}}.}

\item{cveta_out}{If \code{TRUE}, the CV estimate of eta at each point in \code{X} in addition to the CV log-likelihood.}
}
\value{
Scalar value of the cross-validated log-likelihood, or a list with elements \code{eta}, \code{nu} and \code{loglik} if \code{cveta_out = TRUE}.
}
\description{
Leave-one-out local likelihood copula parameter estimates are interpolated, then used to calculate the conditional copula likelihood function.
}
\examples{
# simulate data
family <- 5 # Frank copula
n <- 1000
X <- runif(n) # covariate values
eta_true <- 2*cos(12*pi*X) # copula dependence parameter
par_true <- BiCopEta2Par(family, eta = eta_true)
udata <- VineCopula::BiCopSim(n, family=family,
                              par = par_true$par)

# cross-validation
band <- .1
ieta <- c(1,1)
inu <- 2.5
by <- 10
cv <- CondiCopLikCV(u1= udata[,1], u2= udata[,2],
                    family=family, X=X, eta = ieta, nu = inu,
                    xind = c(seq(1, n, by = by), n), band = band,
                    cveta_out = TRUE)
}
