% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiCopLikCV.R
\name{CondiCopLikCV}
\alias{CondiCopLikCV}
\title{Cross-validated likelihood.}
\usage{
CondiCopLikCV(u1, u2, family, X, xind, degree = c("linear", "constant"),
  eta, nu, kernel = KernEpa, band, cl = NA)
}
\arguments{
\item{u1}{Vector of first uniform response.}

\item{u2}{Vector of second uniform response.}

\item{family}{The copula family to use.  See \code{\link{ConvertPar}}.}

\item{X}{Vector of observed covariate values.}

\item{xind}{Vector of indices in \code{X} at which to calculate leave-one-out parameter estimates.  Can also be supplied as a single integer, in which case \code{xind} equally spaced observations are taken from \code{X}.}

\item{degree}{Character vector specifying degree of local polynomial.  Either "constant" or "linear".}

\item{eta}{Optional initial value of the copula dependence parameter (scalar).}

\item{nu}{Optional initial value of other copula parameters (scalar or vector).}

\item{kernel}{Kernel function.  See \code{\link{kernel}}.}

\item{band}{Bandwidth parameter (positive scalar).}

\item{cl}{Optional parallel cluster created with \code{parallel::makeCluster}.  If \code{NA} runs serially.}
}
\value{
Scalar value of the cross-validated log-likelihood.
}
\description{
Cross-validated likelihood.
}
\details{
Explain exactly what this does...
}
