% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiCopLocFit.R
\name{CondiCopLocFit}
\alias{CondiCopLocFit}
\title{Local likelihood estimation at a single covariate value.}
\usage{
CondiCopLocFit(u1, u2, family, X, x, nx = 100, degree = c("linear",
  "constant"), eta, nu, kernel = KernEpa, band, cl = NA)
}
\arguments{
\item{u1}{Vector of first uniform response.}

\item{u2}{Vector of second uniform response.}

\item{family}{The copula family to use.  See \code{\link{ConvertPar}}.}

\item{X}{Vector of observed covariate values.}

\item{x}{Vector of covariate values within \code{range(X)} at which to fit the local likelihood.  Does not have to be a subset of \code{X}.}

\item{nx}{If \code{x} is missing, defaults to \code{nx} equally spaced values in \code{range(X)}.}

\item{degree}{Character vector specifying degree of local polynomial.  Either "constant" or "linear".}

\item{eta}{Optional initial value of the copula dependence parameter (scalar).}

\item{nu}{Optional initial value of other copula parameters (scalar or vector).}

\item{kernel}{Kernel function.  See \code{\link{kernel}}.}

\item{band}{Bandwidth parameter (positive scalar).}

\item{cl}{Optional parallel cluster created with \code{parallel::makeCluster}.  If \code{NA} runs serially.}
}
\value{
Vector of estimated dependence parameter \code{eta} at each value of \code{x}.
}
\description{
Optionally performs bandwidth and family selection.
}
\note{
FIXME: \code{parallel} requirement.
}
