% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiCopLocFun.R
\name{CondiCopLocFun}
\alias{CondiCopLocFun}
\title{Create a TMB local likelihood function.}
\usage{
CondiCopLocFun(u1, u2, family, X, x, wgt, degree = c("linear",
  "constant"), eta, nu)
}
\arguments{
\item{u1}{Vector of first uniform response.}

\item{u2}{Vector of second uniform response.}

\item{family}{The copula family to use.  See \code{\link{ConvertPar}}.}

\item{X}{Vector of observed covariate values.}

\item{x}{Scalar covariate value at which to evaluate the local likelihood.  Does not have to be a subset of \code{X}.}

\item{wgt}{Vector of kernel weights.}

\item{degree}{Character vector "constant" or "linear" specifying the type of local likelihood fit.}

\item{eta}{Initial value of the copula dependence parameter.}

\item{nu}{Value of other copula parameters (if they exist).}
}
\value{
A list as returned by a call to \code{TMB::MakeADFun}.  In particular, this contains elements \code{fun} and \code{gr} for the *negative* local likelihood and its gradient with respect to \code{eta}.
}
\description{
Wraps a call to \code{TMB::MakeADFun}.
}
