% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernWeights.R
\name{KernWeight}
\alias{KernWeight}
\title{Calculate local likelihood kernel weights.}
\usage{
KernWeight(X, x, band, kernel = KernEpa, band.method = "constant")
}
\arguments{
\item{X}{Vector of covariate values.}

\item{x}{Scalar covariate value at which local likelihood estimation is performed.}

\item{band}{Bandwidth parameter (positive scalar).}

\item{kernel}{Kernel function to use.  See \code{\link{KernFun}}.}

\item{band.method}{Type of bandwith: a character string "constant" or "variable".  See \strong{Details}.}
}
\description{
Calculate local likelihood kernel weights.
}
\details{
For the constant bandwidth of size \code{band = h}, the weights are calculated as
\preformatted{
wgt = kernel((X-x) / h) / h
}
where \code{kernel} is the kernel function.  For bandwidth type "variable", a fixed fraction \code{band} of observations is used, i.e,
\preformatted{
h = sort( abs(X-x) )[ floor(band*length(X)) ]
}
}
