% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernWeights.R
\name{KernWeight}
\alias{KernWeight}
\title{Calculate local likelihood kernel weights.}
\usage{
KernWeight(X, x, band, kernel = KernEpa, band.method = "constant")
}
\arguments{
\item{X}{Vector of covariate values.}

\item{x}{Scalar covariate value at which local likelihood estimation is performed.}

\item{band}{Bandwidth parameter (positive scalar).}
}
\description{
Calculate local likelihood kernel weights.
}
