% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvCopFit.R
\name{SurvCopFit}
\alias{SurvCopFit}
\title{Maximum pseudo-likelihood estimation for survival copulas.}
\usage{
SurvCopFit(u1, u2, status1, status2, family, eta, optim_fun, rm.zero = TRUE)
}
\arguments{
\item{u1}{Vector of first uniform response.}

\item{u2}{Vector of second uniform response.}

\item{status1}{Vector of censoring indicators for the first variable.}

\item{status2}{Vector of censoring indicators for the second variable.}

\item{family}{An integer defining the bivariate copula family to use.  See \code{\link[=ConvertPar]{ConvertPar()}}.}

\item{eta}{Optional initial value of the copula dependence parameter (scalar).  If missing will be estimated unconditionally by \code{\link[VineCopula:BiCopEst]{VineCopula::BiCopEst()}}.}

\item{optim_fun}{Optional specification of likelihood optimization algorithm.  See \strong{Details}.}

\item{rm.zero}{Logical indicating whether zero values in \code{u1} and \code{u2} should be removed when calculating the likelihood.}
}
\value{
\describe{
\item{\code{eta}}{The estimated dependence parameter under the specified copula family.}
}
}
\description{
Estimate the bivariate copula dependence parameter \code{eta} for time-to-event outcomes subject to censoring.
}
\details{
By default, optimization is performed with the quasi-Newton algorithm provided by \code{\link[stats:nlminb]{stats::nlminb()}}, which uses gradient information provided by automatic differentiation (AD) as implemented by \pkg{TMB}.

If the default method is to be overridden, \code{optim_fun} should be provided as a function taking a single argument corresponding to the output of \code{\link[=SurvCopFun]{SurvCopFun()}}, and return a scalar value corresponding to the estimate of \code{eta} at a given covariate value in \code{x0}.  Note that \pkg{TMB} calculates the \emph{negative} local (log)likelihood, such that the objective function is to be minimized.  See \strong{Examples}.
}
\examples{
# the following example shows the local likelihood estimation
# for varying dependence parameter in bivariate survival data

# simulate survival data
set.seed(51)
n <- 1000 # sample size
family <- 3 # Clayton copula
x <- runif(n)
truetau <- function(x, family) {
  eta <- 2 * cos(10 * x^2 - 1) - sin(2 * x^3) + x
  tau <- BiCopEta2Tau(family = family, eta = eta)
  # tau <- rep(runif(1, 0, 1), length(x)) # constant dependence parameter in Kendall's tau scale
  return(tau)
}
plot(x, truetau(x, family), ylim = c(0, 1))
rho <- truetau(x, family)
par <- VineCopula::BiCopTau2Par(family, rho)
marpar1 <- c(1, 1)
marpar2 <- c(2, 2)
marparc <- c(3, 3)
U <- VineCopula::BiCopSim(N = n, family = family, par = par, par2 = 0)
D <- rweibull(n, shape = marparc[2], scale = marparc[1]^(-1 / marparc[2]))
Y <- qweibull(
  U[, 1],
  shape = marpar1[2],
  scale = marpar1[1]^(-1 / marpar1[2]),
  lower.tail = F
)
C <- qweibull(
  U[, 2],
  shape = marpar2[2],
  scale = marpar2[1]^(-1 / marpar2[2]),
  lower.tail = F
)
Y1 <- pmin(Y, D)
Y2 <- pmin(C, D)
status1 <- 1 * (Y <= D)
status2 <- 1 * (C <= D)

# get KM estimates of the survival functions
km1 <- survival::survfit(
  survival::Surv(Y1, status1) ~ 1,
  data = data.frame(Y1, status1)
)
u1 <- km1$surv[match(Y1, km1$time)]
km2 <- survival::survfit(
  survival::Surv(Y2, status2) ~ 1,
  data = data.frame(Y2, status2)
)
u2 <- km2$surv[match(Y2, km2$time)]


# local likelihood estimation
x0 <- seq(min(x), max(x), len = 100)
band <- .1
system.time({
  eta_hat <- SurvCopLocFit(
    u1 = u1,
    u2 = u2,
    status1 = status1,
    status2 = status2,
    family = family,
    x = x,
    x0 = x0,
    band = band
  )
})


# custom optimization routine using stats::optim (gradient-free)
my_optim <- function(obj) {
  opt <- stats::optim(par = obj$par, fn = obj$fn, method = "Nelder-Mead")
  return(opt$par[1]) # always return constant term, even if degree > 0
}
system.time({
  eta_hat2 <- SurvCopLocFit(
    u1 = u1,
    u2 = u2,
    status1 = status1,
    status2 = status2,
    family = family,
    x = x,
    x0 = x0,
    band = band,
    optim_fun = my_optim
  )
})


# compare results
plot(
  x0,
  truetau(x0, family),
  type = "l",
  xlab = expression(x),
  ylab = expression(tau(x)),
  ylim = c(0, 1)
)
lines(x0, BiCopEta2Tau(family, eta = eta_hat$eta), col = "red")
lines(x0, BiCopEta2Tau(family, eta = eta_hat2$eta), col = "blue")
legend(
  "bottomleft",
  fill = c("black", "red", "blue"),
  legend = c("True", "optim_default", "Nelder-Mead")
)
}
