% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvCopLocFit.R
\name{SurvCopLocFit}
\alias{SurvCopLocFit}
\title{Local likelihood estimation for survival copulas.}
\usage{
SurvCopLocFit(
  u1,
  u2,
  status1,
  status2,
  family,
  x,
  x0,
  nx = 100,
  degree = 1,
  eta,
  kernel = KernEpa,
  band,
  optim_fun,
  rm.zero = TRUE,
  cl = NA
)
}
\arguments{
\item{u1}{Vector of first uniform response.}

\item{u2}{Vector of second uniform response.}

\item{status1}{Vector of censoring indicators for the first variable.}

\item{status2}{Vector of censoring indicators for the second variable.}

\item{family}{An integer defining the bivariate copula family to use.  See \code{\link[=ConvertPar]{ConvertPar()}}.}

\item{x}{Vector of observed covariate values.}

\item{x0}{Vector of covariate values within \code{range(x)} at which to fit the local likelihood.  Does not have to be a subset of \code{x}.}

\item{nx}{If \code{x0} is missing, defaults to \code{nx} equally spaced values in \code{range(x)}.}

\item{degree}{Integer specifying the polynomial order of the local likelihood function.  Currently only 0 and 1 are supported.}

\item{eta}{Optional initial value of the copula dependence parameter (scalar).  If missing will be estimated unconditionally by \code{\link[VineCopula:BiCopEst]{VineCopula::BiCopEst()}}.}

\item{kernel}{Kernel function to use.  Should accept a numeric vector parameter and return a non-negative numeric vector of the same length.  See \code{\link[=KernFun]{KernFun()}}.}

\item{band}{Kernal bandwidth parameter (positive scalar).  See \code{\link[=KernWeight]{KernWeight()}}.}

\item{optim_fun}{Optional specification of local likelihood optimization algorithm.  See \strong{Details}.}

\item{rm.zero}{Logical indicating whether zero values in \code{u1} and \code{u2} should be removed when calculating the local likelihood.}

\item{cl}{Optional parallel cluster created with \code{\link[parallel:makeCluster]{parallel::makeCluster()}}, in which case optimization for each element of \code{x0} will be done in parallel on separate cores.  If \code{cl == NA}, computations are run serially.}
}
\value{
List with the following elements:
\describe{
\item{\code{x}}{The vector of covariate values \code{x0} at which the local likelihood is fit.}
\item{\code{eta}}{The vector of estimated dependence parameters of the same length as \code{x0}.}
}
}
\description{
Estimate the bivariate copula dependence parameter \code{eta} at multiple covariate values for time-to-event outcomes subject to censoring.
}
\details{
By default, optimization is performed with the quasi-Newton algorithm provided by \code{\link[stats:nlminb]{stats::nlminb()}}, which uses gradient information provided by automatic differentiation (AD) as implemented by \pkg{TMB}.

If the default method is to be overridden, \code{optim_fun} should be provided as a function taking a single argument corresponding to the output of \code{\link[=SurvCopLocFun]{SurvCopLocFun()}}, and return a scalar value corresponding to the estimate of \code{eta} at a given covariate value in \code{x0}.  Note that \pkg{TMB} calculates the \emph{negative} local (log)likelihood, such that the objective function is to be minimized.  See \strong{Examples}.
}
\examples{
# the following example shows the local likelihood estimation 
# for varying dependence parameter in bivariate survival data

# simulate data
set.seed(31)
n <- 1000 # sample size
family <- 3 # Clayton copula
x <- runif(n)
truetau <- function(x, family){
  eta <- 2*cos(10*x^2-1)-sin(2*x^3)+x
  tau <- BiCopEta2Tau(family = family, eta=eta)
  # tau <- rep(runif(1, 0, 1), length(x)) # constant dependence parameter in Kendall's tau scale
  return(tau)
}
plot(x, truetau(x, family), ylim = c(0,1))
rho <- truetau(x, family)
par <- VineCopula::BiCopTau2Par(family, rho)
sdata <- SurvSim(n, family = family, par = par, 
                 marpar1 = c(1, 1), marpar2 = c(2, 2), marparc = c(3, 3),  
                 mardist = "weibull", type = "rcen")


# get observed data
data <- sdata$data
Y1 <- data[,1]
Y2 <- data[,2]
status1 <- data[,3]
status2 <- data[,4]

# get KM estimates of the survival functions -- for checks only
u1 <- sapply(Y1, function(y){SurvivalCop::KM(t=y, Y = Y1, status = status1)})
u2 <- sapply(Y2, function(y){SurvivalCop::KM(t=y, Y = Y2, status = status2)})


# local likelihood estimation
x0 <- seq(min(x), max(x), len = 100)
band <- .1
system.time({
  eta_hat <- SurvCopLocFit(u1 = u1, u2 = u2,
                           status1 = status1, status2 = status2,
                           family = family, 
                           x = x, x0 = x0, band = band)
})


# custom optimization routine using stats::optim (gradient-free)
my_optim <- function(obj) {
  opt <- stats::optim(par = obj$par, fn = obj$fn, method = "Nelder-Mead")
  return(opt$par[1]) # always return constant term, even if degree > 0
}
system.time({
  eta_hat2 <- SurvCopLocFit(u1 = u1, u2 = u2,
                             status1 = status1, status2 = status2,
                             family = family, 
                             x = x, x0 = x0, band = band,
                             optim_fun = my_optim)
})



# compare results
plot(x0, truetau(x0, family), type = "l", 
     xlab = expression(x), ylab = expression(tau(x)), ylim = c(0,1))
lines(x0, BiCopEta2Tau(family, eta = eta_hat$eta), col = "red")
lines(x0, BiCopEta2Tau(family, eta = eta_hat2$eta), col = "blue")
legend("bottomleft", fill = c("black", "red", "blue"),
       legend = c("True", "optim_default", "Nelder-Mead"))
}
