% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvCopLocFun.R
\name{SurvCopLocFun}
\alias{SurvCopLocFun}
\title{Create a \pkg{TMB} local likelihood function for survival data.}
\usage{
SurvCopLocFun(u1, u2, status1, status2, family, x, x0, wgt, degree = 1, eta)
}
\arguments{
\item{u1}{Vector of first uniform response.}

\item{u2}{Vector of second uniform response.}

\item{status1}{Vector of censoring indicators for the first variable.}

\item{status2}{Vector of censoring indicators for the second variable.}

\item{family}{An integer defining the bivariate copula family to use.  See \code{\link[=ConvertPar]{ConvertPar()}}.}

\item{x}{Vector of observed covariate values.}

\item{x0}{Scalar covariate value at which to evaluate the local likelihood.  Does not have to be a subset of \code{x}.}

\item{wgt}{Vector of positive kernel weights.}

\item{degree}{Integer specifying the polynomial order of the local likelihood function.  Currently only 0 and 1 are supported.}

\item{eta}{Value of the copula dependence parameter.  Scalar or vector of length two, depending on whether \code{degree} is 0 or 1.}
}
\value{
A list as returned by a call to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.  In particular, this contains elements \code{fun} and \code{gr} for the \emph{negative} local likelihood and its gradient with respect to \code{eta}.
}
\description{
Wraps a call to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
